+++ plugins/modules/zhmc_hba.py
@@ -170,11 +170,20 @@ EXAMPLES = """
 """
 
 RETURN = """
+changed:
+  description: Indicates if any change has been made by the module.
+    For C(state=facts), always will be false.
+  returned: always
+  type: bool
+msg:
+  description: An error message that describes the failure.
+  returned: failure
+  type: str
 hba:
   description:
     - "For C(state=absent), an empty dictionary."
-    - "For C(state=present), a dictionary with the resource properties of the
-       HBA after changes, if any."
+    - "For C(state=present), the resource properties of the HBA after any
+       changes."
   returned: success
   type: dict
   contains:
@@ -182,8 +191,10 @@ hba:
       description: "HBA name"
       type: str
     "{property}":
-      description: "Additional properties of the HBA, as described in the
-        :term:`HMC API` (using hyphens (-) in the property names)."
+      description: "Additional properties of the HBA, as described in the data
+        model of the 'HBA' element object of the 'Partition' object in the
+        :term:`HMC API` book.
+        The property names have hyphens (-) as described in that book."
 """
 
 import logging  # noqa: E402